+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 
 - Catch `PaymentProviderAPIException` on CreditCard `post_delete` signal
 - Delete credit card once all orders installments are paid
+- Sorting data in front office is now server-side
 
 ## [2.15.0] - 2025-02-05