+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.16.0] - 2025-02-13
+
 ### Added
 
 - `owner` and `is_main` fields on `CreditCard` model are deprecated
@@ -597,7 +599,8 @@ and this project adheres to
 - First working version serving sellable micro-credentials for multiple
   organizations synchronized to a remote catalog
 
-[unreleased]: https://github.com/openfun/joanie/compare/v2.15.0...main
+[unreleased]: https://github.com/openfun/joanie/compare/v2.16.0...main
+[2.16.0]: https://github.com/openfun/joanie/compare/v2.15.0...v2.16.0
 [2.15.0]: https://github.com/openfun/joanie/compare/v2.14.1...v2.15.0
 [2.14.1]: https://github.com/openfun/joanie/compare/v2.14.0...v2.14.1
 [2.14.0]: https://github.com/openfun/joanie/compare/v2.13.0...v2.14.0